Make the gem work as a rails 3 engine. Made the routes work properly and actually delete the image from database when deleting the attachment

+++ lib/paperclipdb/storage.rb
@@ -10,21 +10,24 @@ module Paperclip
 
       def override_default_options(base)
         @path = @url
-        @url = "/paperclipdb" + @url
+        @url = "/paperclipdb#{@url}"
       end
       private :override_default_options
-      
+
       def exists?(style = default_style)
-        return getAttachment(path(style)).nil?
+        return not(getAttachment(path(style)).nil?)
+      end
+    
+
+      def getAttachment(file_path)
+        file_path = rel_path(file_path)
+        return Paperclipdb::Attachment.find(:first, 
+          :conditions => [ "base_name = ? AND dir_name = ?", File.basename(file_path), File.dirname(file_path) ])
       end
 
-      def getAttachment(file_path) 
-        return Paperclipdb::Attachment.find(:first, :conditions => [ "base_name = ? AND dir_name = ?", File.basename(file_path), File.dirname(file_path) ])
-      end
-      
-       def to_file style = default_style
+      def to_file style = default_style
         if @queued_for_write[style]
-          @queued_for_write[style]
+        @queued_for_write[style]
         elsif exists?(style)
           attachment = getAttachment(path(style))
           tempfile = Tempfile.new attachment.base_name
@@ -34,12 +37,12 @@ module Paperclip
           nil
         end
       end
-  
+
       def flush_writes
         @queued_for_write.each do |style, file|
           attachment = Paperclipdb::Attachment.new
-          attachment.base_name = File.basename(path(style))
-          attachment.dir_name = File.dirname(path(style))
+          attachment.base_name = File.basename(rel_path(path(style)))
+          attachment.dir_name = File.dirname(rel_path(path(style)))
           attachment.content_type = self.instance_variable_get("@_#{self.name.to_s}_content_type")
           attachment.file_size = file.size
           attachment.file_data = file.read
@@ -47,17 +50,18 @@ module Paperclip
         end
         @queued_for_write = {}
       end
-  
+
       def flush_deletes
         @queued_for_delete.each do |path|
-          attachment = getAttachment(path)
-          if (!attachment.nil?)
-            attachment.destroy 
-          end
+          attachment = getAttachment(path)          
+          attachment.destroy if attachment          
         end
         @queued_for_delete = []
-      end        
+      end
       
+      def rel_path path
+        path.to_s.gsub(File.join(Rails.root, 'public'), '')
+      end
     end
   end
 end
